requirements.txt
streamlit
streamlit-chat
transformers
torch
speechbrain
python-dotenv
sendgrid
langchain
langchain-groq